Done with reverting crashing app. Has bugs.
Need more sophisticated copying. Right now it copies the symlinks
recursively, which is not good.

We have done some change of plans here. We can't checkout changes to a
"backup" branch and sync it with backup repo. New plans are on these
lines:

- We copy each changed file to the backup repo
  - make a commit in the backup repo
  - when there is a app crashing change
  - at this point we can copy complete backup to the app dir

We can actually just copy the changed file which is causing the
error back to original repo, I should change code to that. Right now we
copy whole backup dir back to the original repo replacing everything
except hidden files.